feat: user insight 생성 DAG 구현

+++ airflow/sql/fact/fact_user_ratings.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS fact_user_ratings (
     user_id         BIGINT,
-    food_id         INT,
+    food_id         BIGINT,
     food_score      FLOAT,
     created_date    TIMESTAMP, -- 최초 평점 입력 날짜
     PRIMARY KEY (user_id, food_id, created_date),
